+++ backend/canisters/community/impl/src/lifecycle/inspect_message.rs
@@ -19,7 +19,8 @@ fn accept_if_valid(state: &RuntimeState) {
 
     let caller = state.env.caller();
     let is_valid = state.data.members.get(caller).is_some()
-        || (state.data.invited_users.contains(&caller) && method_name == "decline_invitation");
+        // `lookup_user_id` will return true if they are invited to the community
+        || (state.data.members.lookup_user_id(caller).is_some() && method_name == "decline_invitation");
 
     if is_valid {
         ic_cdk::api::call::accept_message();