+++ backend/canisters/group_index/api/src/main.rs
@@ -5,7 +5,6 @@ fn main() {
     generate_candid_method!(group_index, active_groups, query);
     generate_candid_method!(group_index, explore_communities, query);
     generate_candid_method!(group_index, explore_groups, query);
-    generate_candid_method!(group_index, filter_groups, query);
     generate_candid_method!(group_index, recommended_groups, query);
     generate_candid_method!(group_index, search, query);