+++ backend/canisters/group_index/c2c_client/src/lib.rs
@@ -3,7 +3,6 @@ use group_index_canister::*;
 
 // Queries
 generate_c2c_call!(c2c_active_groups);
-generate_c2c_call!(c2c_filter_groups);
 
 // Updates
 generate_c2c_call!(c2c_convert_group_into_community);