+++ backend/canisters/registry/impl/src/lib.rs
@@ -55,15 +55,10 @@ impl RuntimeState {
 #[derive(Serialize, Deserialize)]
 struct Data {
     governance_principals: HashSet<Principal>,
-    #[serde(default = "sns_wasm_canister_id")]
     sns_wasm_canister_id: CanisterId,
     cycles_dispenser_canister_id: CanisterId,
     tokens: Tokens,
     test_mode: bool,
-}
-
-fn sns_wasm_canister_id() -> CanisterId {
-    CanisterId::from_text("qaa6y-5yaaa-aaaaa-aaafa-cai").unwrap()
 }
 
 impl Data {